fix: Improve Widget JS parsing - MEED-8431 - Meeds-io/meeds#2949 (#24)
Prior to this change, when having a missing semicolon or a const in the
JS code, the code isn't executed and an error is thrown for the IDE
Widget in Edit Mode. This change ensures to parse the JS without
introducing a <br> element into it in order to keep the Javascript
engine execute it without strict parsing mode. In addition, this change
will encapsulate the JS code in a specific execution context in order to
allow redefining the same const multiple time in multiple instances in
the same page rather than letting the global context used.